tests/test_models/test_user.py
#!/usr/bin/python3

"""test_User"""
from tests.test_models.test_base_model import test_basemodel
from models.user import User
import os


class test_User(test_basemodel):
    """test_User"""

    def __init__(self, *args, **kwargs):
        """test_User"""
        super().__init__(*args, **kwargs)
        self.name = "User"
        self.value = User

    def test_User_first_name(self):
        """test_User"""
        new = self.value()
        self.assertEqual(type(new.first_name), str if
                         os.getenv('HBNB_TYPE_STORAGE') != 'db' else
                         type(None))

    def test_User_last_name(self):
        """test_User"""
        new = self.value()
        self.assertEqual(type(new.last_name), str if
                         os.getenv('HBNB_TYPE_STORAGE') != 'db' else
                         type(None))

    def test_User_email(self):
        """test_User"""
        new = self.value()
        self.assertEqual(type(new.email), str if
                         os.getenv('HBNB_TYPE_STORAGE') != 'db' else
                         type(None))

    def test_User_password(self):
        """test_User"""
        new = self.value()
        self.assertEqual(type(new.password), str if
                         os.getenv('HBNB_TYPE_STORAGE') != 'db' else
                         type(None))
